chore(cleanup): make notes more understandable
Capitalize the start of lines, make SSH and SCP uppercase, and assume
all commands are run as the normal user (write sudo for commands that
need privelages). I assume that the commands will be run from the normal
user account once it exists. The only exception to this is the bullet
point for setting up sudo access, and on that one I explicitly say that
it has to be done as root.